# expression: implement vectorized evaluation for `builtinTimeToSecSig`