<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

implement vectorized evaluation for builtinTimeToSecSig , for #12103

### What is changed and how it works?

	$go test -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args "builtinTimeToSecSig"
	goos: darwin
	goarch: amd64
	pkg: github.com/pingcap/tidb/expression
	BenchmarkVectorizedBuiltinTimeFuncGenerated-4   	1000000000	         0.0140 ns/op	       0 B/op	       0 allocs/op
	BenchmarkVectorizedBuiltinTimeFunc/builtinTimeToSecSig-VecBuiltinFunc-4         	  100452	     11457 ns/op	       0 B/op	       0 allocs/op
	BenchmarkVectorizedBuiltinTimeFunc/builtinTimeToSecSig-NonVecBuiltinFunc-4      	   40293	     29432 ns/op	       0 B/op	       0 allocs/op
	PASS
	ok  	github.com/pingcap/tidb/expression	2.852s

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test